Updated index.html
I'm all for equal rights for women and men, but this is not the place where I want this discussion to take place. When I use he, you can be sure it's without preference above she, I just wrote how it came to mind. So whenever you encounter he, she's included, when you encounter she, he is included. Women and men are equally precious, so he and she are interchangeable, thus, he = she.
I don't want this topic to surface on this repo again. Thanks for your respect.